+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/conv10_30/Bundle10_30.java
@@ -1,6 +1,6 @@
 package org.hl7.fhir.convertors.conv10_30;
 
-import org.hl7.fhir.convertors.advisor.VersionConvertorAdvisor30;
+import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor30;
 import org.hl7.fhir.convertors.VersionConvertor_10_30;
 import org.hl7.fhir.exceptions.FHIRException;
