+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/conv10_30/ValueSet10_30.java
@@ -1,7 +1,7 @@
 package org.hl7.fhir.convertors.conv10_30;
 
 import java.util.List;
-import org.hl7.fhir.convertors.advisor.VersionConvertorAdvisor30;
+import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor30;
 import org.hl7.fhir.convertors.VersionConvertor_10_30;
 import org.hl7.fhir.dstu2.model.ValueSet;
 import org.hl7.fhir.dstu3.model.CodeSystem;
