+++ org.hl7.fhir.convertors/src/test/java/org/hl7/fhir/convertors/conv10_30/AdministrativeGender10_30Test.java
@@ -1,6 +1,6 @@
 package org.hl7.fhir.convertors.conv10_30;
 
-import org.hl7.fhir.convertors.advisor.VersionConvertorAdvisor30;
+import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor30;
 import org.hl7.fhir.convertors.VersionConvertor_10_30;
 import org.hl7.fhir.convertors.loaders.R2ToR3Loader;
 import org.junit.jupiter.api.DisplayName;
