+++ org.hl7.fhir.convertors/src/test/java/org/hl7/fhir/convertors/conv10_40/Observation10_40Test.java
@@ -1,6 +1,6 @@
 package org.hl7.fhir.convertors.conv10_40;
 
-import org.hl7.fhir.convertors.advisor.VersionConvertorAdvisor40;
+import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor40;
 import org.hl7.fhir.convertors.VersionConvertor_10_40;
 import org.hl7.fhir.convertors.misc.IGR2ConvertorAdvisor;
 import org.junit.jupiter.api.Assertions;
