+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/IgLoader.java
@@ -3,7 +3,7 @@ package org.hl7.fhir.validation;
 import com.google.gson.JsonObject;
 import lombok.Getter;
 import org.hl7.fhir.convertors.*;
-import org.hl7.fhir.convertors.advisor.VersionConvertorAdvisor50;
+import org.hl7.fhir.convertors.advisors.VersionConvertorAdvisor50;
 import org.hl7.fhir.exceptions.FHIRException;
 import org.hl7.fhir.r5.context.SimpleWorkerContext;
 import org.hl7.fhir.r5.elementmodel.Manager;
